Fixes: https://github.com/lcompilers/lpython/issues/2741
Based on: https://github.com/lcompilers/lpython/pull/2805

Tested locally with

```f90
program name
   implicit none
   _lfortran_list(integer):: x
   integer:: i, g
   do i = 1, 100000
      call _lfortran_list_append(x, i)
   end do

   do i = 1, 100000
      g = _lfortran_pop(x, 0)
   end do
end program name
```

Unoptimized
```bash
(lf_20) $ time lfortran test.f90
lfortran test.f90  17.23s user 0.12s system 98% cpu 17.573 total
(lf_20) $ time lfortran test.f90 --fast
lfortran test.f90 --fast  0.91s user 0.12s system 69% cpu 1.488 total
```

Optimized
```bash
(lf_20) $ time lfortran test.f90
lfortran test.f90  0.92s user 0.26s system 36% cpu 3.261 total
(lf_20) $ time lfortran test.f90 --fast
lfortran test.f90 --fast  0.92s user 0.05s system 74% cpu 1.300 total
```

Gives same performance as `--fast` after optimization
17s->0.92s